+++ index.test.js
@@ -1,16 +1,5 @@
 /* eslint-disable max-len */
-const postcss = require('postcss');
-
-const plugin = require('./');
-
-function run(input, output, opts = { }) {
-    return postcss([ plugin(opts) ]).process(input)
-        .then( result => {
-            expect(result.css).toEqual(output);
-            expect(result.warnings().length).toBe(0);
-        });
-}
-
+const run = require('./test-common');
 
 /* Removes !important */
 it('removes !important', () => {
@@ -36,33 +25,4 @@ it('removes behavior, -moz-binding and filter properties', () => {
 it('removes non-GPU transition', () => {
     return run('a { transition: opacity(.5); }; b { transition: skew(.5) };',
         'a { transition: opacity(.5); };', {});
-});
-
-/* Removes disallowed at-rules */
-it('removes @charset', () => {
-    return run('@charset "utf-8";', '', {});
-});
-
-it('removes @import', () => {
-    return run('@import "test.css" screen, projection;', '', {});
-});
-
-it('removes @namespace', () => {
-    return run('@namespace url(http://www.w3.org/1999/xhtml);', '', {});
-});
-
-it('removes @supports', () => {
-    return run('@supports (display: grid) { div { display: grid; } }', '', {});
-});
-
-it('removes @document', () => {
-    return run('@document url("https://www.example.com/") { h1 { color: green; } }', '', {});
-});
-
-it('removes @page', () => {
-    return run('@page { size: a3; }', '', {});
-});
-
-it('removes @viewport', () => {
-    return run('@viewport { width: device-width; }', '', {});
 });
